+++ src/lib/Locale.jsx
@@ -1,0 +1,10 @@
+export const defaultLocale =  {
+    "landingHeaderText": "<questionLength> Questions",
+    "question": "Question",
+    "startQuizBtn": "Start Quiz",
+    "resultFilterAll": "All",
+    "resultFilterCorrect": "Correct",
+    "resultFilterIncorrect": "Incorrect",
+    "nextQuestionBtn": "Next",
+    "resultPageHeaderText": "You have completed the quiz. You got <correctIndexLength> out of <questionLength> questions."
+  }